+++ silo/jsdd.h
@@ -15,6 +15,6 @@ MALLOC size_t* js_dimensions(const json_value* root, size_t* ndims);
 
 /* Figures out the datatype from the JSON parse. */
 enum dtype js_datatype(const json_value* root);
-MALLOC int64_t** js_coord_arrays(const json_value* root, const size_t* dims,
-                                 size_t ndims);
+MALLOC double** js_coord_arrays(const json_value* root, const size_t* dims,
+                                size_t ndims);
 #endif
